Move higher level of !askpatch from admin module to new manager module
